Fix AsyncAPI benchmark bugs and add smoke tests
- Remove ValidationPipelineBenchmarks (redundant with existing schema
  validation benchmarks)
- Remove JsonSchema.Net and NATS.Net package references (unused)
- Fix Utf8JsonWriter state error in PublishAsync with headers: reset
  writer before writing headers after payload
- Fix FormatException in (long)payload.Lumens: Lumens is a double-format
  field, use (double) cast instead
- Add BenchmarkSmokeTests that exercises every benchmark method once
  before BDN starts, catching exceptions immediately rather than after
  a 20+ minute run
- Program.cs runs smoke tests before BDN; supports --smoke-only flag

+++ benchmarks/Corvus.Text.Json.AsyncApi.Benchmarks/RequestReplyBenchmarks.cs
@@ -61,7 +61,7 @@ public class RequestReplyBenchmarks
     }
 
     [Benchmark(Description = "Corvus: Request/Reply (no validation)")]
-    public async ValueTask<long> Corvus_RequestReply_NoValidation()
+    public async ValueTask<double> Corvus_RequestReply_NoValidation()
     {
         (LightMeasuredPayload reply, _) = await this.transport
             .RequestAsync<LightMeasuredPayload, LightMeasuredPayload>(
@@ -70,11 +70,11 @@ public class RequestReplyBenchmarks
                 in this.corvusRequest,
                 CorrelationIdUtf8);
 
-        return (long)reply.Id + (long)reply.Lumens;
+        return (long)reply.Id + (double)reply.Lumens;
     }
 
     [Benchmark(Description = "Corvus: Request/Reply (with basic validation)")]
-    public async ValueTask<long> Corvus_RequestReply_WithValidation()
+    public async ValueTask<double> Corvus_RequestReply_WithValidation()
     {
         // Validate request before sending
         _ = this.corvusRequest.EvaluateSchema();
@@ -88,6 +88,6 @@ public class RequestReplyBenchmarks
 
         // Validate reply
         _ = reply.EvaluateSchema();
-        return (long)reply.Id + (long)reply.Lumens;
+        return (long)reply.Id + (double)reply.Lumens;
     }
 }
